Initial support for pty proxying
To allow (correct) PTY operation in the container,
mount a new instance of devpts on /dev/pts, bind
/dev/pts/ptmx to /dev/ptx and change access rights
on /dev/pts/ptx to 0666. Additionally, the child
process in the container need a new PTY becuse
the parent's PTY is not accessible through the new
devpts instance. We forward data between the
parent's PTY and the child; support for control
characters, etc. (Ctrl-C, Ctrl-D, window size
change, etc.) still needs to be implemented.