Implement attribute translation types
When attribtue emulation is on (-E), file attributes
can be untranslated (passed through "as-is"),
emulated (attributes are stored in .erlent* files),
or mapped (attributes are mapped to the initial user
and/or group ID in the container if they match the user
and/or group ID (or any additional group ID) of the
invoking user).

Command line option for mapped directories is "-M".