Implement an early cleanup (when possible)
After the child has performed its chroot() operation,
we can unmount the fuse filesystem in the parent
(because the child has a separate mount namespace) and
remove the temporary directory. For currently unclear
reasons, this only works when the child has a bind
mount for /dev (otherwise, the temorary directory
is "busy" when trying to unmount it).